refactor(accounts): own account boundary contracts
