+++ src/shared/template-contracts.ts
@@ -1,0 +1,5 @@
+/** The one renderer/main value contract for a game template file. */
+export interface TemplateExportEntry {
+  readonly path: string;
+  readonly contents: string;
+}
